Update `component-scoped-ssr` example
