+++ examples/component-scoped-ssr/src/routes/about/+page.svelte
